+++ modules/viewhelp.sh
@@ -7,28 +7,33 @@ while shell=$(zenity --list --radiolist --width=300 --height=300 --title="$progr
 	arr=$(echo $shell | tr "|" "\n")
 	for x in $arr; do
 		case $x in
-		"repo") if [[ $repotext = "Remove" ]]; then
-			rm -f /etc/yum.repos.d/fedorautils.repo
+		"repo")
+			if [[ $repotext = "Remove" ]]; then
+				rm -f /etc/yum.repos.d/fedorautils.repo
 			else
-			fedorautilsrepo
+				add_repo fedorautils.repo
 			fi
 			detect_repotext;;
-		"issue") show_msg "Opening Browser..."
-				sudo -u "$user"  xdg-open "http://github.com/satya164/fedorautils/issues";;
-		"wiki") show_msg "Opening Browser..."
-				sudo -u "$user"  xdg-open "http://github.com/satya164/fedorautils/wiki";;
-		"changelog") show_msg "Fetching changelog..."
+		"issue")
+			show_msg "Opening Browser..."
+			sudo -u "$user"  xdg-open "http://github.com/satya164/fedorautils/issues";;
+		"wiki")
+			show_msg "Opening Browser..."
+			sudo -u "$user"  xdg-open "http://github.com/satya164/fedorautils/wiki";;
+		"changelog")
+			show_msg "Fetching changelog..."
 			curl -s https://raw.github.com/satya164/fedorautils/master/CHANGELOG -o changelog.txt
 			if [[ -e changelog.txt ]]; then
-			zenity --text-info --width=600 --height=600 --title="Changelog" --filename="changelog.txt" --ok-label="Ok" --cancel-label="Back"
-			rm -f changelog.txt
+				zenity --text-info --width=600 --height=600 --title="Changelog" --filename="changelog.txt" --ok-label="Ok" --cancel-label="Back"
+				rm -f changelog.txt
 			else
-			show_error "Error fetching changelog! Do you have an active internet connection?"
+				show_error "Error fetching changelog! Do you have an active internet connection?"
 			fi;;
-		"logfile") if [[ -e "$logfile" ]]; then
-			zenity --text-info --width=600 --height=600 --title="Logfile" --filename="$logfile" --ok-label="Ok" --cancel-label="Back"
+		"logfile")
+			if [[ -e "$logfile" ]]; then
+				zenity --text-info --width=600 --height=600 --title="Logfile" --filename="$logfile" --ok-label="Ok" --cancel-label="Back"
 			else
-			show_error "No logfile exists. Try running fedorautils with logging enabled. Use --help for more details"
+				show_error "No logfile exists. Try running fedorautils with logging enabled. Use --help for more details"
 			fi;;
 		esac
 	done
@@ -40,21 +45,5 @@ if [[ -e /etc/yum.repos.d/fedorautils.repo ]]; then
 	repotext="Remove"
 else
 	repotext="Add"
-fi
-}
-
-fedorautilsrepo() {
-if [[ -f /etc/yum.repos.d/fedorautils.repo ]]; then
-show_status "Fedora Utils repo already present"
-else
-cat <<EOF | tee /etc/yum.repos.d/fedorautils.repo
-[fedorautils]
-name=Fedora Utils
-baseurl=http://master.dl.sourceforge.net/project/fedorautils/
-enabled=1
-metadata_expire=1d
-skip_if_unavailable=1
-gpgcheck=0
-EOF
 fi
 }
